refactor: shift panels out and text to root
