+++ apps/web/components/user-profile/panels/index.tsx
@@ -1,6 +1,3 @@
 export * from './account'
 export * from './debug'
-export * from './degrees'
-export * from './graphs'
-export * from './modules'
 export * from './public-profile'
